+++ include/lss/eventReactor.hpp
@@ -17,6 +17,7 @@ class EventReactor : public eb::EventHandler<eb::Event>,
                      public eb::EventHandler<StairEvent>,
                      public eb::EventHandler<HeroDiedEvent>,
                      public eb::EventHandler<LocationChangeEvent>,
+                     public eb::EventHandler<PauseEvent>,
                      public eb::EventHandler<QuitCommandEvent> {
 public:
   EventReactor(LSSApp *a) : app(a) {
@@ -32,6 +33,7 @@ public:
     eb::EventBus::AddHandler<StairEvent>(*this);
     eb::EventBus::AddHandler<HeroDiedEvent>(*this);
     eb::EventBus::AddHandler<LocationChangeEvent>(*this);
+    eb::EventBus::AddHandler<PauseEvent>(*this);
   }
   LSSApp *app;
 
@@ -47,6 +49,7 @@ public:
   virtual void onEvent(StairEvent &e) override;
   virtual void onEvent(HeroDiedEvent &e) override;
   virtual void onEvent(LocationChangeEvent &e) override;
+  virtual void onEvent(PauseEvent &e) override;
 
   bool slotCallback(std::shared_ptr<Object>);
   bool itemCallback(std::shared_ptr<Slot>, std::shared_ptr<Object>);