+++ include/lss/game/terrain.hpp
@@ -7,6 +7,10 @@ struct TerrainSpec {
   bool seeThrough;
   bool passThrough;
 
+  bool emitsLight = false;
+  float lightStrength = 0;
+  LightType lightType;
+
   friend bool operator==(TerrainSpec &t1, const TerrainSpec &t2) {
     return t1.name == t2.name;
   }
@@ -24,15 +28,30 @@ public:
     seeThrough = t.seeThrough;
     passThrough = t.passThrough;
     zIndex = 2;
+    emitsLight = t.emitsLight;
+    lightStrength = t.lightStrength;
+    lightType = t.lightType;
+  }
+  Terrain(TerrainSpec t, int z) : Object(), type(t) {
+    seeThrough = t.seeThrough;
+    passThrough = t.passThrough;
+    zIndex = z;
+    emitsLight = t.emitsLight;
+    lightStrength = t.lightStrength;
+    lightType = t.lightType;
   }
   TerrainSpec type;
 };
 
 namespace TerrainType {
-const auto TORCH_STAND = TerrainSpec{"torch stand", false, true};
-const auto STATUE = TerrainSpec{"statue", false, false};
+const auto TORCH_STAND = TerrainSpec{"torch stand",  false,          true, true,
+                                     TORCH_DISTANCE, LightType::FIRE};
+const auto STATUE =
+    TerrainSpec{"statue", false, false, true, 1, LightType::MAGIC};
 const auto ALTAR = TerrainSpec{"altar", true, true};
 const auto BUSH = TerrainSpec{"bush", false, true};
+const auto FIREBALL =
+    TerrainSpec{"fireball", true, true, true, TORCH_DISTANCE, LightType::FIRE};
 }; // namespace TerrainType
 
 #endif // __TERRAIN_H_